feat(uagents-core): resolver interface in uagents-core
