+++ python/uagents-core/uagents_core/utils/resolver.py
@@ -13,6 +13,7 @@ from uagents_core.config import (
 )
 from uagents_core.helpers import weighted_random_sample
 from uagents_core.identity import parse_identifier
+from uagents_core.types import Resolver
 from uagents_core.logger import get_logger
 
 logger = get_logger("uagents_core.utils.resolver")
@@ -71,3 +72,20 @@ def lookup_endpoint_for_agent(
         )
 
     return []
+
+
+class AlmanacResolver(Resolver):
+    def __init__(self, agentverse_config: AgentverseConfig | None = None):
+        self.agentverse_config = agentverse_config or AgentverseConfig()
+
+    async def resolve(self, destination: str) -> tuple[str | None, list[str]]:
+        endpoints = lookup_endpoint_for_agent(
+            agent_identifier=destination, agentverse_config=self.agentverse_config
+        )
+        return None, endpoints
+
+    def sync_resolve(self, destination: str) -> list[str]:
+        endpoints = lookup_endpoint_for_agent(
+            agent_identifier=destination, agentverse_config=self.agentverse_config
+        )
+        return endpoints
